[TL-3] Namespace update
