+++ src/exceptions/SlackBotSenderException.php
@@ -8,6 +8,6 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-namespace Vicky\client\exceptions;
+namespace Vicky\src\exceptions;
 
 class SlackBotSenderException extends \Exception {}
