+++ src/modules/Slack/SlackBotSender.php
@@ -8,9 +8,9 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-namespace Vicky\client\modules\Slack;
+namespace Vicky\src\modules\Slack;
 
-use Vicky\client\exceptions\SlackBotSenderException;
+use Vicky\src\exceptions\SlackBotSenderException;
 
 class SlackBotSender
 {
