+++ src/modules/Vicky.php
@@ -8,7 +8,7 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-namespace Vicky\client\modules;
+namespace Vicky\src\modules;
 
 class Vicky
 {
